abstract ChatBot into smaller components
