+++ src/components/ChatUserMessage.vue
@@ -1,0 +1,14 @@
+<template>
+  <p>{{ message.text }}</p>
+</template>
+
+<script>
+export default {
+  props: {
+    message: {
+      type: Object,
+      required: true
+    }
+  }
+};
+</script>
